# specs/001-resale-pricing-assistant/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Resale Pricing Assistant API
  version: 1.0.0
  description: API v1 para ingesta por foto, propuesta IA y revision humana obligatoria.
servers:
  - url: http://localhost:8000
paths:
  /api/v1/products:
    post:
      summary: Crear producto borrador con fotos
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_channel, photos]
              properties:
                source_channel:
                  type: string
                  enum: [api]
                photos:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [filename, content_base64]
                    properties:
                      filename:
                        type: string
                      content_base64:
                        type: string
      responses:
        '202':
          description: Producto aceptado para analisis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProductResponse'
  /api/v1/proposals/{proposal_id}:
    get:
      summary: Obtener propuesta IA
      operationId: getProposal
      parameters:
        - in: path
          name: proposal_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Propuesta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
        '404':
          description: No encontrada
  /api/v1/proposals/{proposal_id}/review:
    post:
      summary: Aprobar, denegar o editar propuesta
      operationId: reviewProposal
      parameters:
        - in: path
          name: proposal_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '200':
          description: Revision registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '409':
          description: Conflicto de estado o revision concurrente
  /api/v1/products/{product_id}/export:
    post:
      summary: Generar borrador exportable tras aprobacion humana
      operationId: exportApprovedProduct
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Borrador exportable generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '412':
          description: Precondicion incumplida (sin aprobacion humana)
  /api/v1/metrics/llm:
    get:
      summary: Consultar metricas de coste/calidad/fallos LLM
      operationId: getLLMMetrics
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Metricas agregadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMMetricsSummary'
components:
  schemas:
    CreateProductResponse:
      type: object
      required: [product_id, status]
      properties:
        product_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [analyzing]
        proposal_id:
          type: string
          format: uuid
          nullable: true
    Proposal:
      type: object
      required: [proposal_id, product_id, description, suggested_price, confidence_score, status]
      properties:
        proposal_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        description:
          type: string
        suggested_price:
          type: number
        suggested_price_min:
          type: number
          nullable: true
        suggested_price_max:
          type: number
          nullable: true
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
        rationale_internal:
          type: array
          items:
            type: string
        rationale_external:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [proposed, in_review, approved, rejected]
    ReviewRequest:
      type: object
      required: [decision, operator_id]
      properties:
        decision:
          type: string
          enum: [approve, reject, edit]
        operator_id:
          type: string
        edited_description:
          type: string
          nullable: true
        edited_price:
          type: number
          nullable: true
        reject_reason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    ReviewResponse:
      type: object
      required: [review_id, proposal_id, decision, reviewed_at]
      properties:
        review_id:
          type: string
          format: uuid
        proposal_id:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approve, reject, edit]
        reviewed_at:
          type: string
          format: date-time
        next_status:
          type: string
          enum: [approved, rejected, in_review]
    ExportResponse:
      type: object
      required: [product_id, export_id, format, status]
      properties:
        product_id:
          type: string
          format: uuid
        export_id:
          type: string
          format: uuid
        format:
          type: string
          enum: [json, csv, html]
        status:
          type: string
          enum: [ready, exported]
        download_url:
          type: string
          format: uri
    LLMMetricsSummary:
      type: object
      required: [total_calls, total_cost_usd, success_rate, avg_latency_ms]
      properties:
        total_calls:
          type: integer
        total_cost_usd:
          type: number
        success_rate:
          type: number
        avg_latency_ms:
          type: number
        blocked_by_budget:
          type: integer